Moved get_correlation_ids to within Tracer class, changed to be instance method (ignore global tracer)
